requirements.txt
langchain
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
pypdf
unstructured
openpyxl
